more adjustments for testing
Player faces direction of travel, win and lose screens redirect to open screen, new build opens in windowed format

+++ 254 - Codename Tortoise/Assets/Scripts/GlobalObject.cs
@@ -8,7 +8,7 @@ public class GlobalObject : MonoBehaviour
     public static GlobalObject Instance;
     //vars
     public float healthLeft;
-    public float timeRemaining;
+    public double timeRemaining;
 
     private void Awake()
     {
